- Exception en integration test

+++ src/main/backend/databases/controllers/PersonsController.java
@@ -2,6 +2,7 @@ package backend.databases.controllers;
 
 import backend.databases.PersonsDatabase;
 import backend.entry.person.PersonEntry;
+import backend.exception.PersonHasDebtException;
 
 
 import java.util.HashMap;
@@ -22,7 +23,16 @@ public class PersonsController implements Controller<PersonEntry>
 
     @Override
     public void removeEntry(PersonEntry entry) {
-        this.db.removeEntry(entry);
+        entry.calcNetAmount();
+        if (entry.getNetAmount() == 0) {
+            this.db.removeEntry(entry);
+        } else {
+            try {
+                throw new PersonHasDebtException(entry.getName()+" has unsettled debt!");
+            } catch (PersonHasDebtException e) {
+                e.printStackTrace();
+            }
+        }
     }
 
     @Override
